srv/rest-api-srv/requirements.txt
Flask==2.0.1
python-dotenv==0.17.1
grpcio-tools==1.38.0
